==> R/3-Figures-Boxplots.R <==
############################################################################################
############################################################################################
## 3 - BOXPLOT FIGURES COMPARING R-SQUARED VALUES
############################################################################################
############################################################################################
## Returns:
## Boxplot showing distributions of r-squareds for all models and 
## individual r-squareds for the best and worst models of each individual
## based on the mean r-squared

### Packages ----
libs <- c('data.table', 'ggplot2', 'cowplot', 'dplyr', 'DescTools')
lapply(libs, require, character.only = TRUE)

### Load elkyears with fewer than 750 points ----
# Npoints_750<- readRDS('input/exclusion_elk.rds')

### Load data from RSF outputs folder ----
all_correlations <- readRDS('results/prediction_outputs/correlations.rds')


# Calculate difference in all combinations of model comparisons for each iteration
# Individual & ranef model
MedianCI(all_correlations$ranef_corr, conf.level=0.95)
# Individual & GFR model
MedianCI(all_correlations$gfr_corr, conf.level=0.95)
# Individual & No FR
MedianCI(all_correlations$WS_corr, conf.level=0.95)

# Calculate mean correlations for all individuals
meanByID <- data.table()
for (i in unique(all_correlations$elkyear)){
  sub_corrs <- subset(all_correlations, elkyear==i)
  
  for(model in c('WS_corr', 'ranef_corr', 'gfr_corr')){
    sub_model <- unlist(sub_corrs %>% select(model))
    if(length(sub_model)<6){
      next
    }
    assign(paste(model, 'mean', sep='_'), signif(mean(sub_model),3))
    assign(paste(model, 'lower', sep='_'), signif(as.numeric(MeanCI(sub_model, conf.level=0.95)[2]),3))
    assign(paste(model, 'upper', sep='_'), signif(as.numeric(MeanCI(sub_model, conf.level=0.95)[3]),3))
  }
  
  summaries <- data.table(elkyear=rep(unlist(i),3), model=c('WS_corr', 'ranef_corr', 'gfr_corr'), mean=c(WS_corr_mean, ranef_corr_mean, gfr_corr_mean), 
                          lower=c(WS_corr_lower, ranef_corr_lower, gfr_corr_lower), upper=c(WS_corr_upper, ranef_corr_upper, gfr_corr_upper))
  meanByID <- rbind(meanByID, summaries)
}

# Pick the best and worst model for each individual
bestmodelscores <- data.table()
worstmodelscores <- data.table()
for (i in unique(meanByID$elkyear)){
  indiv_sub <- meanByID[elkyear==i]
  indiv_sub <- indiv_sub[1:3,]
  best_mod <- indiv_sub[mean==max(indiv_sub$mean)]
  worst_mod <- indiv_sub[mean==min(indiv_sub$mean)]
  best <- data.table(elkyear=unlist(best_mod$elkyear), best=unlist(best_mod$model), best_score=unlist(best_mod$mean), lower=unlist(best_mod$lower), upper=unlist(best_mod$upper))
  worst <- data.table(elkyear=unlist(worst_mod$elkyear), worst=unlist(worst_mod$model), worst_score=unlist(worst_mod$mean), lower=unlist(worst_mod$lower), upper=unlist(worst_mod$upper))
  bestmodelscores <- rbind(bestmodelscores, best)
  worstmodelscores <- rbind(worstmodelscores, worst)
}

# Make data.table for supplementary material
supp_correlations <- merge(bestmodelscores[,1:3], worstmodelscores[,1:3], by='elkyear')
supp_correlations <- supp_correlations[order(elkyear)]

# Make data.table for boxplot
melt_correlations <- melt(all_correlations, measure.vars=c('WS_corr', 'ranef_corr', 'gfr_corr'))

# Make data.table for individual correlations
corr_indivs <- melt_correlations %>% group_by(elkyear, variable) %>% summarize(mean=mean(value, na.rm=T), 
                                                                               lower=mean(value, na.rm=T)-(sd(value, na.rm=T)/sqrt(length(value))*1.96),
                                                                               upper=mean(value, na.rm=T)+(sd(value, na.rm=T)/sqrt(length(value))*1.96))

# tiff('figures/fig2.tiff', width = 7, height = 6, units = 'in', res = 300)

ggplot(melt_correlations, aes(x=variable, y=value, col=variable)) + 
  # Add zero line at median of individual level model
  geom_hline(yintercept=median(melt_correlations[variable=='individual']$value), col='#000000', linetype='dashed') +
  # Add boxplots
  geom_boxplot(notch=TRUE, aes(fill=variable), width=0.4, colour='black', alpha=0.5) + 
  # Set boxpot colours
  scale_fill_manual(values=c("#65ABF3", "#DED9DC", "#DED9DC", "#DED9DC")) + 
  scale_x_discrete(labels=c('individual' = 'Individual', 'no_FR' = 'Base',
                            'ranef' = 'Ran. Eff.', 'gfr' = 'GFR')) +
  # Theme
  theme(panel.background = element_rect(fill='white'), panel.border = element_rect(colour='black', fill=NA), 
        legend.position='none', axis.text = element_text(size=12, colour='black'), axis.title.y = element_text(size=15, colour='black', vjust=2.5)) +
  ylab(expression(paste('Correlation with individual level model (R'^2, ')'))) + xlab('') +
  # Optionally add points to show worst/best models
  # geom_point(data=worstmodelscores, aes(x=worst, y=worst_score), col='black', fill='white', position=position_jitter(width=.1), pch=21, alpha=0.5) +
  # geom_point(data=bestmodelscores, aes(x=best, y=best_score), col='black', fill='black', position=position_jitter(width=.1), pch=24, alpha=0.5)\
  # Add individuals
  geom_point(data=corr_indivs, aes(x=variable, y=mean, group=elkyear), position=position_dodge(width=0.5)) +
  geom_errorbar(data=corr_indivs, aes(x=variable, y=mean, group=elkyear, ymin=lower, ymax=upper), width=2, position=position_dodge(width=0.5)) +
  geom_smooth(data=corr_indivs, aes(x=variable, y=mean, group=elkyear), se=F, method='lm', colour='#C0C0C075')
  
# Save mean correlations by elk year
saveRDS(meanByID, 'input/mean_correlations.rds')

# Save table of scores for supplementary material
write.csv(supp_correlations, 'tables/supplementary_model_scores.csv')
